affiliate/be-affiliate/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic
python-multipart
python-jose
cryptograpy
passlib[bcrypt]
gunicorn